does not download if file has exists

+++ edinet_xbrl/edinet_xbrl_downloader.py
@@ -10,6 +10,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License. See accompanying LICENSE file.
 
+import os
 from edinet_xbrl.ufocatcher_util import UfoCatcherUtil
 from time import sleep
 import urllib.request
@@ -19,11 +20,14 @@ class EdinetXbrlDownloader(object):
   @staticmethod
   def download(url, target_dir):
     file_name = "{0}/{1}".format(target_dir, url.split("/")[-1])
-    urllib.request.urlretrieve(url, file_name)
+    if not os.path.exists(file_name):
+        urllib.request.urlretrieve(url, file_name)
+        return True
+    return False
 
   @classmethod
   def download_by_ticker(cls, ticker, target_dir, wait_sec=1.0):
     response = UfoCatcherUtil.request(ticker)
     for url in UfoCatcherUtil.generate_edinet_xbrl_url(response.text):
-      cls.download(url, target_dir)
-      sleep(wait_sec)
+        if cls.download(url, target_dir):
+              sleep(wait_sec)
